requirements.txt
python-dotenv==1.0.1
spacy==4.0.0.dev3
genai==2.1.0
langchain-core==0.3.51
sentence-transformers==4.1.0
langgraph==0.3.28
langchain-neo4j==0.4.0
